chore: move management helpers out of __init__.py
It is better to collect these in a standalone module.
