+++ weblate/runner.py
@@ -10,7 +10,7 @@ from contextlib import suppress
 def main(argv=None, developer_mode: bool = False) -> None:
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "weblate.settings")
 
-    from weblate.utils.management import WeblateManagementUtility
+    from weblate.utils.management.utility import WeblateManagementUtility
 
     if argv is None:
         argv = sys.argv
