+++ weblate/trans/management/commands/commit_pending.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from weblate.trans.management.commands import WeblateLangCommand
 from weblate.trans.tasks import commit_pending
+from weblate.utils.management.base import WeblateLangCommand
 
 if TYPE_CHECKING:
     from django.core.management.base import CommandParser
