+++ weblate/trans/management/commands/updategit.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from weblate.trans.management.commands import WeblateComponentCommand
 from weblate.trans.tasks import perform_update
+from weblate.utils.management.base import WeblateComponentCommand
 
 if TYPE_CHECKING:
     from django.core.management.base import CommandParser
